double circular queue with two policies

+++ src/disjoint.cc
@@ -2,12 +2,12 @@
 
 int disjoint_set::find(int i) const
 {
-    if (auto next_root = storage_.at(i); next_root < 0) 
+    if (auto next_root = storage_.at(static_cast<size_t>(i)); next_root < 0) 
         return i;
     else return find(next_root);
 }
 
 void disjoint_set::union_sets(int root1, int root2)
 {
-    storage_.at(root2) = root1;
+    storage_.at(static_cast<size_t>(root2)) = root1;
 }
